requirements.txt
giotto_tda==0.6.0
matplotlib==3.7.1
networkx==3.1
numpy==1.23.5
pandas==2.0.3
persim==0.3.1
plotly==5.14.1
point_cloud_utils==0.29.6
pyvista==0.39.0
ripser==0.6.4
scikit_learn==1.3.0
scipy==1.9.1
skimage==0.0
tadasets==0.0.4
umap_learn==0.5.3
